+++ src/components/accordion/AccordionItem.tsx
@@ -1,0 +1,25 @@
+import { Item } from '@radix-ui/react-accordion'
+import React from 'react'
+
+import { styled } from '~/stitches'
+
+const StyledItem = styled(Item, {
+  width: '100%',
+  overflow: 'hidden',
+
+  '&:not(:last-child)': {
+    mb: '$1'
+  }
+})
+
+type AccordionItemProps = React.ComponentProps<typeof StyledItem>
+
+export const AccordionItem: React.FC<AccordionItemProps> = ({
+  children,
+  value,
+  ...remainingProps
+}) => (
+  <StyledItem value={value} {...remainingProps}>
+    {children}
+  </StyledItem>
+)