+++ src/components/action-icon/ActionIcon.tsx
@@ -1,8 +1,9 @@
 import type * as Stitches from '@stitches/react'
 import invariant from 'invariant'
+import { darken } from 'polished'
 import * as React from 'react'
 
-import { styled } from '~/stitches'
+import { styled, theme } from '~/stitches'
 import { NavigatorActions } from '~/types'
 import { Override } from '~/utilities'
 
@@ -82,7 +83,8 @@ const StyledButton = styled('button', {
     },
     size: {
       md: { size: '$3' },
-      lg: { size: '$4' }
+      lg: { size: '$4' },
+      xl: { size: '$5' }
     },
     isRounded: {
       true: {
@@ -101,6 +103,15 @@ const StyledButton = styled('button', {
       theme: 'primary',
       appearance: 'simple',
       css: getSimpleVariant('$primary', '$primaryMid', '$primaryDark')
+    },
+    {
+      theme: 'primaryDark',
+      appearance: 'simple',
+      css: getSimpleVariant(
+        '$primaryDark',
+        darken(0.1, theme.colors.primaryDark.value),
+        darken(0.15, theme.colors.primaryDark.value)
+      )
     },
     {
       theme: 'success',
@@ -125,6 +136,15 @@ const StyledButton = styled('button', {
       css: getSolidVariant('$primary', '$primaryMid', '$primaryDark')
     },
     {
+      theme: 'primaryDark',
+      appearance: 'solid',
+      css: getSolidVariant(
+        '$primaryDark',
+        darken(0.1, theme.colors.primaryDark.value),
+        darken(0.15, theme.colors.primaryDark.value)
+      )
+    },
+    {
       theme: 'success',
       appearance: 'solid',
       css: getSolidVariant('$success', '$successMid', '$successDark')
@@ -145,6 +165,15 @@ const StyledButton = styled('button', {
       theme: 'primary',
       appearance: 'outline',
       css: getOutlineVariant('$primary', '$primaryMid', '$primaryDark')
+    },
+    {
+      theme: 'primaryDark',
+      appearance: 'outline',
+      css: getOutlineVariant(
+        '$primaryDark',
+        darken(0.1, theme.colors.primaryDark.value),
+        darken(0.15, theme.colors.primaryDark.value)
+      )
     },
     {
       theme: 'success',
@@ -225,7 +254,7 @@ export const ActionIcon = React.forwardRef<HTMLButtonElement, ActionIconProps>(
 
           return React.cloneElement(child, {
             css: {
-              size: size === 'lg' ? 20 : 16,
+              size: ['lg', 'xl'].includes(size as string) ? 20 : 16,
               ...(child.props.css ? child.props.css : {})
             }
           })