+++ src/components/table/TableHeader.tsx
@@ -1,5 +1,29 @@
 import { styled } from '~/stitches'
+import React from 'react'
+import { TableHeaderCell } from './TableHeaderCell'
 
-export const TableHeader = styled('thead', {})
+const StyledTableHeader = styled('thead', {
+  variants: {
+    theme: {
+      primary: {
+        [`${TableHeaderCell}`]: {
+          bg: '$primary'
+        }
+      },
+      primaryDark: {
+        [`${TableHeaderCell}`]: {
+          bg: '$primaryDark'
+        }
+      }
+    }
+  }
+})
+
+type TableHeaderProps = React.ComponentProps<typeof StyledTableHeader>
+
+export const TableHeader: React.FC<TableHeaderProps> = ({
+  theme = 'primaryDark',
+  ...rest
+}: TableHeaderProps) => <StyledTableHeader theme={theme} {...rest} />
 
 TableHeader.displayName = 'TableHeader'