+++ src/components/table/TableHeaderCell.tsx
@@ -1,22 +1,13 @@
 import { styled } from '~/stitches'
 
 export const TableHeaderCell = styled('th', {
-  bg: '$primaryDark',
   color: 'white',
   fontFamily: '$body',
   fontWeight: 600,
   lineHeight: 1.5,
   p: '$2 $3',
   textAlign: 'left',
-  verticalAlign: 'middle',
-  '&:first-of-type': {
-    borderTopLeftRadius: '$0',
-    borderBottomLeftRadius: '$0'
-  },
-  '&:last-of-type': {
-    borderTopRightRadius: '$0',
-    borderBottomRightRadius: '$0'
-  }
+  verticalAlign: 'middle'
 })
 
 TableHeaderCell.displayName = 'TableHeaderCell'