+++ src/components/tabs/Tabs.test.tsx
@@ -19,9 +19,32 @@ const TabsTest = ({ defaultValue = 'tab1' }) => (
   </IdProvider>
 )
 
+const MobileTabsTest = ({ defaultValue = 'tab1' }) => (
+  <IdProvider>
+    <Tabs defaultValue={defaultValue}>
+      <Tabs.TriggerList enableTabScrolling>
+        <Tabs.Trigger value="tab1">
+          Trigger 1 which is going to be long
+        </Tabs.Trigger>
+        <Tabs.Trigger value="tab2">
+          Trigger 2 which is going to be even longer
+        </Tabs.Trigger>
+      </Tabs.TriggerList>
+      <Tabs.Content value="tab1">Important content for tab 1</Tabs.Content>
+      <Tabs.Content value="tab2">Important content for tab 2</Tabs.Content>
+    </Tabs>
+  </IdProvider>
+)
+
 describe('Tabs component', () => {
   it('renders', async () => {
     const { container } = render(<TabsTest />)
+
+    expect(container).toMatchSnapshot()
+  })
+
+  it('renders mobile view', async () => {
+    const { container } = render(<MobileTabsTest />)
 
     expect(container).toMatchSnapshot()
   })